+++ sdk/schemaregistry/azure-data-schemaregistry/src/test/java/com/azure/data/schemaregistry/SchemaRegistryAsyncClientTest.java
@@ -16,14 +16,16 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Mono;
+import reactor.test.StepVerifier;
 
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Objects;
 import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.function.Function;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -40,7 +42,8 @@ public class SchemaRegistryAsyncClientTest {
     private static final SchemaId MOCK_SCHEMA_ID = new SchemaId();
     private static final String MOCK_GROUP = "mockgroup";
     private static final String MOCK_SCHEMA_NAME = "mockname";
-    private static final String MOCK_AVRO_SCHEMA = "{\"namespace\":\"example2.avro\",\"type\":\"record\",\"name\":\"User\",\"fields\":[{\"name\":\"name\",\"type\":\"string\"},{\"name\":\"favorite_number\",\"type\": [\"int\", \"null\"]}]}";
+    private static final String MOCK_AVRO_SCHEMA =
+        "{\"namespace\":\"example2.avro\",\"type\":\"record\",\"name\":\"User\",\"fields\":[{\"name\":\"name\",\"type\":\"string\"},{\"name\":\"favorite_number\",\"type\": [\"int\", \"null\"]}]}";
 
     private SchemaRegistryAsyncClient client;
     private AzureSchemaRegistry restService;
@@ -51,8 +54,8 @@ public class SchemaRegistryAsyncClientTest {
 
     @BeforeEach
     protected void setUp() {
-        this.guidCache = new HashMap<String, SchemaProperties>();
-        this.schemaStringCache = new HashMap<String, SchemaProperties>();
+        this.guidCache = new HashMap<>();
+        this.schemaStringCache = new HashMap<>();
 
         this.typeParserDictionary = new ConcurrentSkipListMap<>(String.CASE_INSENSITIVE_ORDER);
         this.typeParserDictionary.put(MOCK_SERIALIZATION.toString(), (s) -> s);
@@ -71,7 +74,7 @@ public class SchemaRegistryAsyncClientTest {
     }
 
     @Test
-    public void testRegisterThenSchemaCacheHit() throws Exception {
+    public void testRegisterThenSchemaCacheHit() {
         MOCK_SCHEMA_ID.setId(MOCK_ID);
         when(restService.getSchemas()).thenReturn(schemas);
         when(schemas.registerWithResponseAsync(anyString(), anyString(),
@@ -86,21 +89,20 @@ public class SchemaRegistryAsyncClientTest {
                         MOCK_SCHEMA_ID,
                         null)));
 
-        assertEquals(
-            MOCK_ID,
-            client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION)
-                .block().getSchemaId());
-        assertEquals(
-            MOCK_ID,
-            client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION)
-                .block().getSchemaId());
+        StepVerifier.create(client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(properties -> assertEquals(MOCK_ID, properties.getSchemaId()))
+            .verifyComplete();
+
+        StepVerifier.create(client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(properties -> assertEquals(MOCK_ID, properties.getSchemaId()))
+            .verifyComplete();
 
         verify(schemas, times(1))
             .registerWithResponseAsync(MOCK_GROUP, MOCK_SCHEMA_NAME, SerializationType.AVRO, MOCK_AVRO_SCHEMA);
     }
 
     @Test
-    public void testGetGuidThenSchemaCacheHit() throws Exception {
+    public void testGetGuidThenSchemaCacheHit() {
         MOCK_SCHEMA_ID.setId(MOCK_ID);
         when(restService.getSchemas()).thenReturn(schemas);
         when(schemas.queryIdByContentWithResponseAsync(anyString(), anyString(),
@@ -114,17 +116,20 @@ public class SchemaRegistryAsyncClientTest {
                         MOCK_SCHEMA_ID,
                         null)));
 
-        assertEquals(MOCK_ID,
-            client.getSchemaId(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION).block());
-        assertEquals(MOCK_ID,
-            client.getSchemaId(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION).block());
+        StepVerifier.create(client.getSchemaId(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(schemaId -> assertEquals(MOCK_ID, schemaId))
+            .verifyComplete();
+
+        StepVerifier.create(client.getSchemaId(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(schemaId -> assertEquals(MOCK_ID, schemaId))
+            .verifyComplete();
 
         verify(schemas, times(1))
             .queryIdByContentWithResponseAsync(anyString(), anyString(), any(SerializationType.class), anyString());
     }
 
     @Test
-    public void testGetSchemaThenGuidCacheHit() throws Exception {
+    public void testGetSchemaThenGuidCacheHit() {
         String mockId = "mock-id---";
         SchemasGetByIdHeaders mockHeaders = new SchemasGetByIdHeaders();
         mockHeaders.setXSchemaType(MOCK_SERIALIZATION.toString());
@@ -138,13 +143,26 @@ public class SchemaRegistryAsyncClientTest {
                     MOCK_AVRO_SCHEMA,
                     mockHeaders)));
 
-        SchemaProperties first = client.getSchema(mockId.toString()).block();
-        SchemaProperties second = client.getSchema(mockId.toString()).block();
-
-        assertTrue(first.equals(second));
-        assertEquals(mockId.toString(), first.getSchemaId());
+        StepVerifier.create(client.getSchema(mockId)
+            .flatMap(properties -> Mono.just(properties).zipWith(client.getSchema(mockId))))
+            .assertNext(tuple2 -> {
+                assertEquals(mockId, tuple2.getT1().getSchemaId());
+                assertTrue(areSchemaPropertiesEqual(tuple2.getT1(), tuple2.getT2()));
+            })
+            .verifyComplete();
 
         verify(schemas, times(1)).getByIdWithResponseAsync(mockId);
+    }
+
+    private static boolean areSchemaPropertiesEqual(SchemaProperties properties1, SchemaProperties properties2) {
+        if (properties1 == null) {
+            return properties2 == null;
+        }
+
+        return Arrays.equals(properties1.getSchema(), properties2.getSchema())
+            && Objects.equals(properties1.getSchemaId(), properties2.getSchemaId())
+            && Objects.equals(properties1.getSchemaName(), properties2.getSchemaName())
+            && Objects.equals(properties1.getSerializationType(), properties2.getSerializationType());
     }
 
     @Test
@@ -162,10 +180,9 @@ public class SchemaRegistryAsyncClientTest {
                         MOCK_SCHEMA_ID,
                         null)));
 
-        assertEquals(
-            MOCK_ID,
-            client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION)
-                .block().getSchemaId());
+        StepVerifier.create(client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(properties -> assertEquals(MOCK_ID, properties.getSchemaId()))
+            .verifyComplete();
 
         client.clearCache();
 
@@ -175,10 +192,9 @@ public class SchemaRegistryAsyncClientTest {
 
         this.typeParserDictionary.put(MOCK_SERIALIZATION.toString(), (s) -> s);
 
-        assertEquals(
-            MOCK_ID,
-            client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION)
-                .block().getSchemaId());
+        StepVerifier.create(client.registerSchema(MOCK_GROUP, MOCK_SCHEMA_NAME, MOCK_AVRO_SCHEMA, MOCK_SERIALIZATION))
+            .assertNext(properties -> assertEquals(MOCK_ID, properties.getSchemaId()))
+            .verifyComplete();
 
         verify(schemas, times(2))
             .registerWithResponseAsync(MOCK_GROUP, MOCK_SCHEMA_NAME, SerializationType.AVRO, MOCK_AVRO_SCHEMA);
@@ -198,16 +214,10 @@ public class SchemaRegistryAsyncClientTest {
                         null,
                         null,
                         null)));
-        try {
-            client.registerSchema(
-                "doesn't matter",
-                "doesn't matter",
-                "doesn't matter",
-                MOCK_SERIALIZATION).block();
-            fail("Should throw on 400 status code");
-        } catch (IllegalStateException e) {
-            assert true;
-        }
+
+        String doesntMatter = "doesn't matter";
+        StepVerifier.create(client.registerSchema(doesntMatter, doesntMatter, doesntMatter, MOCK_SERIALIZATION))
+            .verifyError(IllegalStateException.class);
 
         verify(schemas, times(1))
             .registerWithResponseAsync(anyString(), anyString(), any(SerializationType.class), anyString());
@@ -228,16 +238,9 @@ public class SchemaRegistryAsyncClientTest {
                         null,
                         null)));
 
-        try {
-            client.getSchemaId(
-                "doesn't matter",
-                "doesn't matter",
-                "doesn't matter",
-                MOCK_SERIALIZATION).block();
-            fail("Should throw on 404 status code");
-        } catch (IllegalStateException e) {
-            assert true;
-        }
+        String doesntMatter = "doesn't matter";
+        StepVerifier.create(client.getSchemaId(doesntMatter, doesntMatter, doesntMatter, MOCK_SERIALIZATION))
+            .verifyError(IllegalStateException.class);
 
         verify(schemas, times(1))
             .queryIdByContentWithResponseAsync(anyString(), anyString(), any(SerializationType.class), anyString());
@@ -256,14 +259,8 @@ public class SchemaRegistryAsyncClientTest {
                     null,
                     null)));
 
-        try {
-            client.getSchema(mockId).block();
-            fail("Should have thrown on 404 status code");
-        } catch (IllegalStateException e) {
-            assert true;
-        } catch (Exception e) {
-            assert false;
-        }
+        StepVerifier.create(client.getSchema(mockId))
+            .verifyError(IllegalStateException.class);
 
         verify(schemas, times(1))
             .getByIdWithResponseAsync(mockId);