+++ sdk/spring/azure-spring-boot-test-aad/src/test/java/com/azure/test/aad/groups/AADGroupsCountIT.java
@@ -3,8 +3,8 @@
 
 package com.azure.test.aad.groups;
 
-import com.microsoft.azure.spring.autoconfigure.aad.AADAuthenticationFilter;
-import com.microsoft.azure.spring.autoconfigure.aad.UserPrincipal;
+import com.azure.spring.autoconfigure.aad.AADAuthenticationFilter;
+import com.azure.spring.autoconfigure.aad.UserPrincipal;
 import com.azure.test.oauth.OAuthResponse;
 import com.azure.test.oauth.OAuthUtils;
 import com.azure.test.utils.AppRunner;
@@ -26,8 +26,8 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import static com.azure.test.oauth.OAuthUtils.AAD_SINGLE_TENANT_CLIENT_ID;
 import static com.azure.test.oauth.OAuthUtils.AAD_SINGLE_TENANT_CLIENT_SECRET;
@@ -83,8 +83,8 @@ public class AADGroupsCountIT {
             String groupsCount = Optional.of(authToken)
                                          .map(PreAuthenticatedAuthenticationToken::getPrincipal)
                                          .map(p -> (UserPrincipal) p)
-                                         .map(UserPrincipal::getUserGroups)
-                                         .map(List::size)
+                                         .map(UserPrincipal::getGroups)
+                                         .map(Set::size)
                                          .map(String::valueOf)
                                          .orElse("");
             return new ResponseEntity<>(groupsCount, HttpStatus.OK);